pyproject.toml
[tool.poetry]
name = "thmanyah-assessment"
version = "0.1.0"
description = ""
authors = ["yousef_vm <yoosif@windowslive.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.25"
kafka-python = "^2.0.2"
requests = "^2.31.0"
pandas = "^2.1.4"
pyarrow = "^15.0.0"
psycopg2-binary = "^2.9.9"
tqdm = "^4.66.2"
python-dotenv = "^1.0.0"
clickhouse-connect = "^0.8.18"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
